+++ packages/isolate-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ricsam/isolate-client
 
+## 0.1.3
+
+### Patch Changes
+
+- add eval timeout and async iterator for custom functions
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes