+++ packages/isolate-protocol/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ricsam/isolate-protocol
 
+## 0.1.3
+
+### Patch Changes
+
+- add eval timeout and async iterator for custom functions
+
 ## 0.1.2
 
 ### Patch Changes